+++ tools/find_unused_blocks.py
@@ -7,23 +7,18 @@ from pathlib import Path
 BASE_DXF_FOLDER = r'D:\source\dxftest'
 
 
-def _get_all_block_references(dwg):
-    block_references = []
-    for layout in dwg.layouts_and_blocks():
-        block_references.extend(layout.query('INSERT'))
-    return block_references
-
-
 def _find_unused_blocks(dwg):
-    references = _get_all_block_references(dwg)
-    used_block_names = set(entity.dxf.name for entity in references)
-    existing_blocks = set(block.name for block in dwg.blocks if not block.is_layout_block)
-    unused_blocks = existing_blocks - used_block_names
+    references = dwg.query('INSERT')
+    # block names are case insensitive!
+    used_block_names = set(entity.dxf.name.lower() for entity in references)
+    # exclude layout blocks, they are not referenced by any DXF entity
+    existing_blocks = set(block.name.lower() for block in dwg.blocks if not block.is_layout_block)
+    not_referenced_blocks = existing_blocks - used_block_names
     references_without_definition = used_block_names - existing_blocks
 
-    if len(unused_blocks):
+    if len(not_referenced_blocks):
         print('\nFound unused BLOCK definitions:\n')
-        print('\n'.join(sorted(unused_blocks)))
+        print('\n'.join(sorted(not_referenced_blocks)))
 
     if len(references_without_definition):
         print('\nFound BLOCK references without definition:\n')